Move flags test to PostHogFeatureFlagsV3Test.kt
